Create helper for initializing tests and JSDOM
