+++ assets/js/vectorscores.js
@@ -4,7 +4,8 @@ layout: compress-js
 
 var VS = {};
 
-{% include_relative _helpers.js %}
+{% include_relative _helpers.util.js %}
+{% include_relative _helpers.math.js %}
 {% include_relative _layout.js %}
 {% include_relative _score.js %}
 {% include_relative _controls.js %}
